LPS-89371 portal-search-test-util: move common test classes from legacy portal-test
https://issues.liferay.com/browse/LPS-89371
