+++ modules/apps/portal-search/portal-search-test-util/src/main/java/com/liferay/portal/search/test/util/SearchTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search.test;
+package com.liferay.portal.search.test.util;
 
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentHelper;
